TVSBACKEND/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==2.3.0
django==4.2.23
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-otp==1.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
packaging==25.0
pillow==10.4.0
propcache==0.2.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
twilio==9.6.3
typing-extensions==4.13.2
urllib3==2.2.3
yarl==1.15.2
